refactor: select user for update when balance is being updated

+++ migrations/000002_create_transactions_table.up.sql
@@ -3,7 +3,7 @@ CREATE TYPE transaction_source_type AS ENUM ('game', 'server', 'payment');
 
 CREATE TABLE transactions (
     transaction_id UUID PRIMARY KEY,
-    user_id INTEGER NOT NULL REFERENCES users(id) ON DELETE CASCADE,
+    user_id BIGINT NOT NULL REFERENCES users(id) ON DELETE CASCADE,
     amount DECIMAL(12, 2) NOT NULL,
     state transaction_state NOT NULL,
     source_type transaction_source_type NOT NULL,
